+++ src/common/auto-checker.ts
@@ -2,6 +2,6 @@
 // Licensed under the MIT License.
 import { A11YSelfValidator } from './a11y-self-validator';
 
-declare interface AutoChecker {
+export interface AutoChecker {
     A11YSelfValidator: A11YSelfValidator;
 }
